Create basic-node-calculator
A little project I did which uses a library called readline-sync. 

I had to use the following command

Run npm install --save readline-sync

And then run the necessary code at the top of the JS file to tell Node.js that my app needs to load the library so that I can use the prompt functionality.  That allowed me to use the console.log to retrieve user inputs.

The user input is returned as a string so I had to watch out for converting to and integer where appropriate.